+++ tests/websocket/close-dropped-server/src/api/middleware-trigger-drop/meta.ts
@@ -6,7 +6,9 @@ export const middleware: Middleware[] = [
   async function (req, _res, next) {
     const body = await req.json() as { clientId: string }
 
-    req.ws.drop(KickedCloseSignal, id => id === body.clientId)
+    req.ws.drop(KickedCloseSignal, session =>
+      session.clientId === body.clientId,
+    )
 
     return next()
   },