+++ tests/websocket/server-query/src/api/handler-trigger-query/get.ts
@@ -5,18 +5,16 @@ type Query = {
 }
 
 type ConnectionData = {
-  app: {
-    userId: string
-  }
+  userId: string
 }
 
 export default async function handler (req: Request): AsyncHandlerResult {
   const { userId } = req.query as Query
 
-  const entries = req.ws.query((_clientId, data) => {
-    const connectionData = data as ConnectionData
+  const entries = req.ws.query((session) => {
+    const data = session.data as ConnectionData
 
-    return connectionData.app.userId !== userId
+    return data.userId !== userId
   })
 
   return Response.json({